fix: address code review findings across 8 files
Bugs:
- evaluation.py: add missing upload_int import (would NameError at runtime)
- initialization.py: remove redundant first astype(float32) in he_init
- optim/sgd.py: remove broad except Exception swallowing real training errors

Security:
- flex/trainer.py: torch.load weights_only=False -> weights_only=True

Correctness / reliability:
- checkpoints.py: atomic checkpoint saves via tmp + os.replace to prevent
  corrupt files on crash mid-write
- cuda_backend.py: double-checked locking on get_lib() to prevent race
  condition during concurrent library initialization

Refactoring (train_cuda.py):
- Replace 43 individual `from settings import` with `import settings as S`
  to reduce import noise and simplify future additions
- Extract _conv_backward_step() helper to deduplicate the four nearly-
  identical conv backward blocks (conv1-conv4)
- Remove module-level globals (d_w_conv1 etc.) and module-level helper
  functions that referenced them; use closure-based _device_weights() /
  _velocity_buffers() locals within main() instead
- Fix misleading batch progress log: show per-batch acc, not running acc

Minor:
- flex/builder.py: replace manual product loop with math.prod()

+++ src/minicnn/models/initialization.py
@@ -7,7 +7,7 @@ from minicnn.config.settings import C1_IN, C1_OUT, C2_IN, C2_OUT, C3_IN, C3_OUT,
 
 def he_init(size, fan_in, rng=None):
     rng = rng or np.random.default_rng()
-    return (rng.standard_normal(size).astype(np.float32) * np.sqrt(2.0 / fan_in)).astype(np.float32)
+    return (rng.standard_normal(size) * np.sqrt(2.0 / fan_in)).astype(np.float32)
 
 
 def init_weights(seed):
